requirements.txt
Django==4.1
django-enumfield==3.1
django-tinymce==3.6.1
django-tinymce4-lite==1.8.0
func-timeout==4.3.5
jsmin==3.0.1
mistune==0.8.4
numpy==1.24.3
Pillow==9.5.0
sqlparse==0.4.4
tzdata==2023.3
django-ckeditor
diff-match-patch
mysqlclient==2.2.0
python-dotenv
django-cron
PyJWT
django-embed-video
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.7.2  # Valinnainen, tarpeen mukaan
pycryptodome
